Inject either hosted fields or card fields depending on context
